requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
filelock==3.20.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
griffe==1.15.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.12
llama-index-cli==0.5.3
llama-index-core==0.14.12
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-groq==0.4.1
llama-index-llms-openai==0.6.13
llama-index-llms-openai-like==0.5.3
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-tools-tavily-research==0.4.2
llama-index-workflows==2.12.2
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
openai==2.15.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==6.33.4
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
streamlit==1.53.0
striprtf==0.0.26
tavily-python==0.7.19
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wrapt==1.17.3
yarl==1.22.0